requirements.txt
fastapi[standard]>=0.115
pydantic>=2.10
pydantic-settings>=2.6
python-telegram-bot>=21
httpx>=0.27
eval_type_backport>=0.2  # let pydantic evaluate `X | None` syntax on py3.9
